+++ history.md
@@ -4,6 +4,10 @@
 
 ### AzGovViz version 5
 
+__Changes__ (2021-Sep-01 / Major)
+
+* Update AzAPICall CostManagement return
+
 __Changes__ (2021-Aug-30 / Major)
 
 * Adding feature for RBAC Role assignments: determine 'standing' from PIM (Priviledged Identity Mangement) managed Role assignments